+++ src/main/java/de/adorsys/keycloak/config/repository/RealmRepository.java
@@ -23,6 +23,7 @@ package de.adorsys.keycloak.config.repository;
 import de.adorsys.keycloak.config.exception.KeycloakRepositoryException;
 import de.adorsys.keycloak.config.provider.KeycloakProvider;
 import de.adorsys.keycloak.config.util.ResponseUtil;
+import org.apache.commons.lang3.ObjectUtils;
 import org.keycloak.admin.client.Keycloak;
 import org.keycloak.admin.client.resource.RealmResource;
 import org.keycloak.admin.client.resource.RealmsResource;
@@ -31,7 +32,10 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.stereotype.Service;
 
+
 import java.util.List;
+import java.util.HashMap;
+
 
 import jakarta.ws.rs.NotFoundException;
 import jakarta.ws.rs.WebApplicationException;
@@ -61,7 +65,10 @@ public class RealmRepository {
     }
 
     public RealmRepresentation get(String realmName) {
-        return getResource(realmName).toRepresentation();
+        final var realm = getResource(realmName).toRepresentation();
+        realm.setAttributes(ObjectUtils.firstNonNull(realm.getAttributes(), new HashMap<>()));
+        realm.setEventsEnabled(ObjectUtils.firstNonNull(realm.isEventsEnabled(), false));
+        return realm;
     }
 
     public void create(RealmRepresentation realm) {