+++ src/test/java/de/adorsys/keycloak/config/service/ImportAuthenticationFlowsIT.java
@@ -761,11 +761,15 @@ class ImportAuthenticationFlowsIT extends AbstractImportIT {
 
         List<AuthenticatorConfigRepresentation> authConfig;
         authConfig = getAuthenticatorConfig(realm, "id1");
-        assertThat(authConfig, hasSize(2));
+        int expectedAuthConfigSize = VersionUtil.lt(KEYCLOAK_VERSION, "26.1")? 2 : 1;
+
+        assertThat(authConfig, hasSize(expectedAuthConfigSize));
         assertThat(authConfig.get(0).getAlias(), is("id1"));
         assertThat(authConfig.get(0).getConfig(), hasEntry(is("defaultProvider"), is("id1")));
-        assertThat(authConfig.get(1).getAlias(), is("id1"));
-        assertThat(authConfig.get(1).getConfig(), hasEntry(is("defaultProvider"), is("id1")));
+        if (VersionUtil.lt(KEYCLOAK_VERSION, "26.1")) {
+            assertThat(authConfig.get(1).getAlias(), is("id1"));
+            assertThat(authConfig.get(1).getConfig(), hasEntry(is("defaultProvider"), is("id1")));
+        }
 
         authConfig = getAuthenticatorConfig(realm, "id2");
         assertThat(authConfig, hasSize(1));
@@ -792,11 +796,15 @@ class ImportAuthenticationFlowsIT extends AbstractImportIT {
 
         List<AuthenticatorConfigRepresentation> authConfig;
         authConfig = getAuthenticatorConfig(realm, "id1");
-        assertThat(authConfig, hasSize(2));
+        int expectedAuthConfigSize = VersionUtil.lt(KEYCLOAK_VERSION, "26.1")? 2 : 1;
+
+        assertThat(authConfig, hasSize(expectedAuthConfigSize));
         assertThat(authConfig.get(0).getAlias(), is("id1"));
         assertThat(authConfig.get(0).getConfig(), hasEntry(is("defaultProvider"), is("id2")));
-        assertThat(authConfig.get(1).getAlias(), is("id1"));
-        assertThat(authConfig.get(1).getConfig(), hasEntry(is("defaultProvider"), is("id2")));
+        if (VersionUtil.lt(KEYCLOAK_VERSION, "26.1")) {
+            assertThat(authConfig.get(1).getAlias(), is("id1"));
+            assertThat(authConfig.get(1).getConfig(), hasEntry(is("defaultProvider"), is("id2")));
+        }
 
         authConfig = getAuthenticatorConfig(realm, "id2");
         assertThat(authConfig, hasSize(1));